pyproject.toml
[tool.poetry]
name = "movie_finder_django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "n"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4"
annotated-types = "==0.7.0"
anthropic = "==0.31.2"
anyio = "==4.4.0"
asgiref = "==3.8.1"
certifi = "==2024.7.4"
cffi = "==1.17.1"
charset-normalizer = "==3.3.2"
cryptography = "==44.0.2"
distro = "==1.9.0"
django = "==5.0.7"
django-allauth = "==65.4.1"
django-cors-headers = "==4.4.0"
django-environ = "==0.12.0"
django-extensions = "==3.2.3"
django-filter = "==24.3"
djangorestframework = "==3.15.2"
djangorestframework-simplejwt = "==5.3.1"
filelock = "==3.15.4"
fsspec = "==2024.6.1"
h11 = "==0.14.0"
httpcore = "==1.0.5"
httpx = "==0.27.0"
huggingface-hub = "==0.24.0"
idna = "==3.7"
iniconfig = "==2.1.0"
jiter = "==0.5.0"
packaging = "==24.1"
pluggy = "==1.5.0"
psycopg2-binary = "==2.9.10"
pycparser = "==2.22"
pydantic = "==2.8.2"
pydantic-core = "==2.20.1"
pyjwt = "==2.8.0"
python-dotenv = "==1.0.1"
pyyaml = "==6.0.1"
requests = "==2.32.3"
requests-mock = "==1.12.1"
setuptools = "==75.6.0"
sniffio = "==1.3.1"
sqlparse = "==0.5.0"
tokenizers = "==0.19.1"
tqdm = "==4.66.4"
typing-extensions = "==4.12.2"
ua-parser = "==1.0.0"
ua-parser-builtins = "==0.18.0.post1"
urllib3 = "==2.2.2"
user-agents = "==2.2.0"
wheel = "==0.45.1"
pytest = "==8.3.5"
drf-spectacular = "^0.28.0"

[tool.poetry.dev-dependencies]
pytest-django = "==4.10.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
ruff = "^0.13.0"

[tool.ruff]
line-length = 150
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "DJ"]
ignore = ["DJ008"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[build-system]
requires = ["poetry-core>=1.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"
